Fix handling of GNOME background color
- The GNOME shell service doesn't handle 48bits values that can be set in
the /desktop/gnome/background/primary_color gconf key (which are valid and
properly supported by Gtk+). In this case, the "Set as desktop background"
functionality is broken by the exception that is returned by the GNOME shell
service:
 Error: uncaught exception: [Exception... "Component returned failure code:
 0x80004005 (NS_ERROR_FAILURE) [nsIShellService.desktopBackgroundColor]"
 nsresult: "0x80004005 (NS_ERROR_FAILURE)"  location: "JS frame ::
 chrome://browser/content/setDesktopBackground.js :: anonymous :: line 56"
 data: no]

- When the gconf key is 24 or 12 bits, the values GetDesktopBackgroundColor()
returns are broken, and the background color displayed in the "Set as desktop
background" dialog is (almost) black. The problem is it only keeps the 4 last
bits for each R,G,B value.

- HexToNum() doesn't return the correct decimal value for hexa values from
0xa to 0xf.